requirements.txt
asgiref==3.7.2
Django==5.0.1
django-address==0.2.8
django-jazzmin==2.6.0
django-phonenumber-field==7.3.0
phonenumbers==8.13.29
pillow==10.2.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
tzdata==2023.4
